+++ src/main/java/ca/maximilian/cursed_craft/mixin/CrashOnDeath.java
@@ -2,6 +2,7 @@ package ca.maximilian.cursed_craft.mixin;
 
 import ca.maximilian.cursed_craft.Config;
 import ca.maximilian.cursed_craft.CrashOnDeathType;
+import com.sun.jdi.ThreadReference;
 import net.minecraft.client.Minecraft;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.damagesource.DamageSource;
@@ -13,20 +14,27 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(ServerPlayer.class)
 public class CrashOnDeath {
-    @Inject(method="die", at=@At("HEAD"))
+    @Inject(method="die", at=@At("RETURN"))
     private void onDeath(DamageSource source, CallbackInfo info) {
-        if (Config.HANDLER.instance().crashOnDeathType == CrashOnDeathType.MEMSET) {
-            MemoryUtil.memSet(0, 0, 1L);
-        } else if (Config.HANDLER.instance().crashOnDeathType == CrashOnDeathType.STOPPEDRESPONDING) {
-            Minecraft.getInstance().execute(() -> {
-                try {
-                    while(true) {
-                        Thread.sleep(100000L);
+        new Thread(() -> {
+            try {
+                Thread.sleep(200);
+            } catch (InterruptedException e) {
+                throw new RuntimeException(e);
+            }
+            if (Config.HANDLER.instance().crashOnDeathType == CrashOnDeathType.MEMSET) {
+                MemoryUtil.memSet(0, 0, 1L);
+            } else if (Config.HANDLER.instance().crashOnDeathType == CrashOnDeathType.STOPPEDRESPONDING) {
+                Minecraft.getInstance().execute(() -> {
+                    try {
+                        while(true) {
+                            Thread.sleep(100000L);
+                        }
+                    } catch (InterruptedException e) {
+                        throw new RuntimeException(e);
                     }
-                } catch (InterruptedException e) {
-                    throw new RuntimeException(e);
-                }
-            });
-        }
+                });
+            }
+        }).start();
     }
 }